@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(autoware_health_checker)
 
-add_compile_options(-std=c++11)
+add_compile_options(-std=c++14)
 
 find_package(catkin REQUIRED COMPONENTS
   autoware_system_msgs
@@ -10,6 +10,7 @@ find_package(catkin REQUIRED COMPONENTS
   jsk_rviz_plugins
   rosunit
   rosgraph_msgs
+  roslint
 )
 
 catkin_package(
@@ -18,6 +19,9 @@ catkin_package(
   CATKIN_DEPENDS autoware_system_msgs rosgraph_msgs
 #  DEPENDS system_lib
 )
+
+set(ROSLINT_CPP_OPTS "--filter=-build/c++11")
+roslint_cpp()
 
 include_directories(
   include
@@ -28,6 +32,8 @@ set(HEALTH_CHECKER_SRC
   src/health_checker/health_checker.cpp
   src/health_checker/diag_buffer.cpp
   src/health_checker/rate_checker.cpp
+  src/health_checker/value_manager.cpp
+  src/health_checker/param_manager.cpp
 )
 add_library(health_checker
   ${HEALTH_CHECKER_SRC}
@@ -45,7 +51,7 @@ add_executable(health_aggregator
   src/health_aggregator/health_aggregator_node.cpp
   src/health_aggregator/health_aggregator.cpp
 )
-target_link_libraries(health_aggregator ${catkin_LIBRARIES})
+target_link_libraries(health_aggregator health_checker ${catkin_LIBRARIES})
 add_dependencies(health_aggregator ${catkin_EXPORTED_TARGETS})
 
 add_executable(health_analyzer
@@ -75,6 +81,16 @@ install(TARGETS health_checker system_status_subscriber
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
 
+# Install launch
+install(DIRECTORY launch/
+  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch
+)
+
+# Install config
+install(DIRECTORY config/
+  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/config
+)
+
 if (CATKIN_ENABLE_TESTING)
   find_package(rostest REQUIRED)
   add_rostest_gtest(test-autoware_health_checker
@@ -83,4 +99,6 @@ if (CATKIN_ENABLE_TESTING)
   ${HEALTH_CHECKER_SRC})
   target_link_libraries(test-autoware_health_checker
   ${catkin_LIBRARIES})
+
+  roslint_add_test()
 endif ()
