@@ -17,230 +17,255 @@
  * v1.0 Masaya Kataoka
  */
 
+#include <string>
+#include <vector>
 #include <autoware_health_checker/health_checker/health_checker.h>
 
-namespace autoware_health_checker {
-HealthChecker::HealthChecker(ros::NodeHandle nh,
-                                         ros::NodeHandle pnh) {
-  node_activated_ = false;
-  ros_ok_ = true;
-  nh_ = nh;
-  pnh_ = pnh;
+namespace autoware_health_checker
+{
+HealthChecker::HealthChecker(ros::NodeHandle nh, ros::NodeHandle pnh)
+  : value_manager_(nh, pnh)
+  , node_activated_(false)
+  , nh_(nh)
+  , pnh_(pnh)
+{
   status_pub_ =
-      nh_.advertise<autoware_system_msgs::NodeStatus>("node_status", 10);
+    nh_.advertise<autoware_system_msgs::NodeStatus>("node_status", 10);
 }
 
-HealthChecker::~HealthChecker() {
-  ros_ok_ = false;
-  publish_thread_.join();
-}
-
-void HealthChecker::publishStatus() {
-  ros::Rate rate = ros::Rate(autoware_health_checker::UPDATE_RATE);
-  while (ros_ok_) {
-    mtx_.lock();
-    autoware_system_msgs::NodeStatus status;
-    status.node_activated = node_activated_;
-    ros::Time now = ros::Time::now();
-    status.header.stamp = now;
-    status.node_name = ros::this_node::getName();
-    std::vector<std::string> checker_keys = getRateCheckerKeys();
-    // iterate Rate checker and publish rate_check result
-    for (auto key_itr = checker_keys.begin(); key_itr != checker_keys.end();
-         key_itr++) {
-      autoware_system_msgs::DiagnosticStatusArray diag_array;
-      autoware_system_msgs::DiagnosticStatus diag;
-      diag.type = autoware_system_msgs::DiagnosticStatus::UNEXPECTED_RATE;
-      std::pair<uint8_t, double> result =
-          rate_checkers_[*key_itr]->getErrorLevelAndRate();
-      diag.level = result.first;
-      diag.key = *key_itr;
-      diag.value = valueToJson(result.second);
-      diag.description = rate_checkers_[*key_itr]->description;
+void HealthChecker::publishStatus(const ros::TimerEvent& event)
+{
+  std::lock_guard<std::mutex> lock(mtx_);
+  autoware_system_msgs::NodeStatus status;
+  status.node_activated = node_activated_;
+  ros::Time now = ros::Time::now();
+  status.header.stamp = now;
+  status.node_name = ros::this_node::getName();
+  const auto checker_keys = getRateCheckerKeys();
+  // iterate Rate checker and publish rate_check result
+  for (const auto& key : checker_keys)
+  {
+    const auto result = rate_checkers_[key]->getErrorLevelAndRate();
+    if (result)
+    {
+      AwDiagStatusArray diag_array;
+      AwDiagStatus diag = setValueCommon(
+        key, result->second, rate_checkers_.at(key)->description);
       diag.header.stamp = now;
+      diag.level = result->first;
+      diag.type = AwDiagStatus::UNEXPECTED_RATE;
       diag_array.status.push_back(diag);
       status.status.push_back(diag_array);
     }
-    // iterate Diagnostic Buffer and publish all diagnostic data
-    std::vector<std::string> keys = getKeys();
-    for (auto key_itr = keys.begin(); key_itr != keys.end(); key_itr++) {
-      status.status.push_back(diag_buffers_[*key_itr]->getAndClearData());
-    }
-    status_pub_.publish(status);
-    mtx_.unlock();
-    rate.sleep();
   }
-  return;
-}
-
-uint8_t HealthChecker::SET_DIAG_STATUS(autoware_system_msgs::DiagnosticStatus status) {
-  if (status.level == LEVEL_OK || status.level == LEVEL_WARN ||
-    status.level == LEVEL_ERROR || status.level == LEVEL_FATAL) {
-    addNewBuffer(status.key, status.type, status.description);
-    diag_buffers_[status.key]->addDiag(status);
-    return status.level;
-  } else {
-    return LEVEL_UNDEFINED;
+  // iterate Diagnostic Buffer and publish all diagnostic data
+  const auto keys = getKeys();
+  for (const auto& key : keys)
+  {
+    status.status.push_back(diag_buffers_[key]->getAndClearData());
   }
+  status_pub_.publish(status);
 }
 
-uint8_t HealthChecker::CHECK_TRUE(std::string key, bool value,
-  uint8_t level, std::string description) {
-  if (level == LEVEL_OK || level == LEVEL_WARN ||
-    level == LEVEL_ERROR || level == LEVEL_FATAL) {
-    autoware_system_msgs::DiagnosticStatus status;
-    status.key = key;
-    status.value = valueToJson(value);
-    status.level = level;
-    status.header.stamp = ros::Time::now();
-    addNewBuffer(key, status.INVALID_VALUE, description);
-    diag_buffers_[status.key]->addDiag(status);
-    return level;
-  } else {
-    return LEVEL_UNDEFINED;
+ErrorLevel HealthChecker::SET_DIAG_STATUS(
+  const autoware_system_msgs::DiagnosticStatus& status)
+{
+  value_manager_.addCandidate(status.key);
+  if (value_manager_.isNotFound(status.key))
+  {
+    return AwDiagStatus::UNDEFINED;
   }
+  auto identify = [&status](const ErrorLevel target_level)
+  {
+    return (status.level == target_level);
+  };
+  if (!identify(AwDiagStatus::OK) && !identify(AwDiagStatus::WARN) &&
+    !identify(AwDiagStatus::ERROR) && !identify(AwDiagStatus::FATAL))
+  {
+    return AwDiagStatus::UNDEFINED;
+  }
+  addNewBuffer(status.key, status.type, status.description);
+  return status.level;
 }
 
-void HealthChecker::ENABLE() {
-  publish_thread_ = boost::thread(boost::bind(&HealthChecker::publishStatus, this));
-  return;
+ErrorLevel HealthChecker::CHECK_TRUE(
+  const ErrorKey& key, const bool value,
+  const ErrorLevel level, const std::string& description)
+{
+  value_manager_.addCandidate(key);
+  if (value_manager_.isNotFound(key))
+  {
+    return AwDiagStatus::UNDEFINED;
+  }
+  AwDiagStatus status = setValueCommon(key, value, description);
+  status.level = level;
+  status.type = status.INVALID_VALUE;
+  return SET_DIAG_STATUS(status);
 }
 
-std::vector<std::string> HealthChecker::getKeys() {
-  std::vector<std::string> keys;
-  std::vector<std::string> checker_keys = getRateCheckerKeys();
-  for (auto buf_itr = diag_buffers_.begin();
-       buf_itr != diag_buffers_.end(); buf_itr++) {
-    bool matched = false;
-    for (auto checker_key_itr = checker_keys.begin();
-         checker_key_itr != checker_keys.end(); checker_key_itr++) {
-      if (*checker_key_itr == buf_itr->first) {
-        matched = true;
-      }
-    }
-    if (!matched) {
-      keys.push_back(buf_itr->first);
+void HealthChecker::ENABLE()
+{
+  ros::Duration duration(1.0 / autoware_health_checker::UPDATE_RATE);
+  timer_ = nh_.createTimer(duration, &HealthChecker::publishStatus, this);
+}
+
+std::vector<ErrorKey> HealthChecker::getKeys()
+{
+  std::vector<ErrorKey> keys;
+  const auto checked = getRateCheckerKeys();
+  for (const auto& buf : diag_buffers_)
+  {
+    const auto res = std::find(checked.begin(), checked.end(), buf.first);
+    if (res == checked.end())
+    {
+      keys.push_back(buf.first);
     }
   }
   return keys;
 }
 
-std::vector<std::string> HealthChecker::getRateCheckerKeys() {
-  std::vector<std::string> keys;
-  for (auto itr = rate_checkers_.begin(); itr != rate_checkers_.end(); itr++) {
-    keys.push_back(itr->first);
+std::vector<ErrorKey> HealthChecker::getRateCheckerKeys()
+{
+  std::vector<ErrorKey> keys;
+  for (const auto& checker : rate_checkers_)
+  {
+    keys.push_back(checker.first);
   }
   return keys;
 }
 
-bool HealthChecker::keyExist(std::string key) {
-  if (diag_buffers_.count(key) == 0) {
-    return false;
-  }
-  return true;
+bool HealthChecker::keyExist(const ErrorKey& key) const
+{
+  return (diag_buffers_.count(key) != 0);
 }
 
 // add New Diagnostic Buffer if the key does not exist
-bool HealthChecker::addNewBuffer(std::string key, uint8_t type, std::string description) {
-  mtx_.lock();
-  bool is_success = false;
-  if (!keyExist(key)) {
-    std::unique_ptr<DiagBuffer> buf_ptr(
-      new DiagBuffer(key, type, description, autoware_health_checker::BUFFER_LENGTH));
-    diag_buffers_[key] = std::move(buf_ptr);
-    is_success = true;
+bool HealthChecker::addNewBuffer(
+  const ErrorKey& key, const ErrorType type, const std::string& description)
+{
+  std::lock_guard<std::mutex> lock(mtx_);
+  if (keyExist(key))
+  {
+    return false;
   }
-  mtx_.unlock();
-  return is_success;
+  diag_buffers_[key] = std::make_unique<DiagBuffer>(key, type, description,
+    autoware_health_checker::BUFFER_DURATION);
+  return true;
 }
 
-uint8_t HealthChecker::CHECK_MIN_VALUE(std::string key, double value,
-                                             double warn_value,
-                                             double error_value,
-                                             double fatal_value,
-                                             std::string description) {
-  addNewBuffer(key, autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE,
-               description);
-  autoware_system_msgs::DiagnosticStatus new_status;
-  new_status.type = autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE;
-  if (value < fatal_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::FATAL;
-  } else if (value < error_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::ERROR;
-  } else if (value < warn_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::WARN;
-  } else {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::OK;
+ErrorLevel HealthChecker::CHECK_MIN_VALUE(const ErrorKey& key,
+  const double value, const double warn_value, const double error_value,
+  const double fatal_value, const std::string& description)
+{
+  value_manager_.addCandidate(key);
+  if (value_manager_.isNotFound(key))
+  {
+    return AwDiagStatus::UNDEFINED;
   }
-  new_status.description = description;
-  new_status.value = valueToJson(value);
-  new_status.header.stamp = ros::Time::now();
-  diag_buffers_[key]->addDiag(new_status);
-  return new_status.level;
+  static const ThreshType thresh_type = "min";
+  value_manager_.setDefaultValue(
+    key, thresh_type, warn_value, error_value, fatal_value);
+  auto identify = [key, value, thresh_type, this](ErrorLevel level)
+  {
+    return (value < value_manager_.getValue(key, thresh_type, level).get());
+  };
+  AwDiagStatus new_status = setValueCommon(key, value, description);
+  new_status.level = identify(AwDiagStatus::FATAL) ? AwDiagStatus::FATAL :
+    identify(AwDiagStatus::ERROR) ? AwDiagStatus::ERROR :
+    identify(AwDiagStatus::WARN) ? AwDiagStatus::WARN :
+    AwDiagStatus::OK;
+  new_status.type = AwDiagStatus::OUT_OF_RANGE;
+  return SET_DIAG_STATUS(new_status);
 }
 
-uint8_t HealthChecker::CHECK_MAX_VALUE(std::string key, double value,
-                                             double warn_value,
-                                             double error_value,
-                                             double fatal_value,
-                                             std::string description) {
-  addNewBuffer(key, autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE,
-               description);
-  autoware_system_msgs::DiagnosticStatus new_status;
-  new_status.type = autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE;
-  if (value > fatal_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::FATAL;
-  } else if (value > error_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::ERROR;
-  } else if (value > warn_value) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::WARN;
-  } else {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::OK;
+ErrorLevel HealthChecker::CHECK_MAX_VALUE(const ErrorKey& key,
+  const double value, const double warn_value, const double error_value,
+  const double fatal_value, const std::string& description)
+{
+  value_manager_.addCandidate(key);
+  if (value_manager_.isNotFound(key))
+  {
+    return AwDiagStatus::UNDEFINED;
   }
-  new_status.description = description;
-  new_status.value = valueToJson(value);
-  new_status.header.stamp = ros::Time::now();
-  diag_buffers_[key]->addDiag(new_status);
-  return new_status.level;
+  static const ThreshType thresh_type = "max";
+  value_manager_.setDefaultValue(
+    key, thresh_type, warn_value, error_value, fatal_value);
+  auto identify = [key, value, thresh_type, this](ErrorLevel level)
+  {
+    return (value > value_manager_.getValue(key, thresh_type, level).get());
+  };
+  AwDiagStatus new_status = setValueCommon(key, value, description);
+  new_status.level = identify(AwDiagStatus::FATAL) ? AwDiagStatus::FATAL :
+    identify(AwDiagStatus::ERROR) ? AwDiagStatus::ERROR :
+    identify(AwDiagStatus::WARN) ? AwDiagStatus::WARN :
+    AwDiagStatus::OK;
+  new_status.type = AwDiagStatus::OUT_OF_RANGE;
+  return SET_DIAG_STATUS(new_status);
 }
 
-uint8_t HealthChecker::CHECK_RANGE(std::string key, double value,
-                                         std::pair<double, double> warn_value,
-                                         std::pair<double, double> error_value,
-                                         std::pair<double, double> fatal_value,
-                                         std::string description) {
-  addNewBuffer(key, autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE,
-               description);
-  autoware_system_msgs::DiagnosticStatus new_status;
-  new_status.type = autoware_system_msgs::DiagnosticStatus::OUT_OF_RANGE;
-  if (value < fatal_value.first || value > fatal_value.second) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::FATAL;
-  } else if (value < error_value.first || value > error_value.second) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::ERROR;
-  } else if (value < warn_value.first || value > warn_value.second) {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::WARN;
-  } else {
-    new_status.level = autoware_system_msgs::DiagnosticStatus::OK;
+ErrorLevel HealthChecker::CHECK_RANGE(const ErrorKey& key,
+  const double value, const MinMax warn_value, const MinMax error_value,
+  const MinMax fatal_value, const std::string& description)
+{
+  value_manager_.addCandidate(key);
+  if (value_manager_.isNotFound(key))
+  {
+    return AwDiagStatus::UNDEFINED;
   }
-  new_status.value = valueToJson(value);
-  new_status.description = description;
-  new_status.header.stamp = ros::Time::now();
-  diag_buffers_[key]->addDiag(new_status);
-  return new_status.level;
+  value_manager_.setDefaultValue(key, "min", warn_value.first,
+    error_value.first, fatal_value.first);
+  value_manager_.setDefaultValue(key, "max", warn_value.second,
+    error_value.second, fatal_value.second);
+  auto identify = [key, value, this](ErrorLevel level)
+  {
+    return (value < value_manager_.getValue(key, "min", level).get()) ||
+      (value > value_manager_.getValue(key, "max", level).get());
+  };
+  AwDiagStatus new_status = setValueCommon(key, value, description);
+  new_status.level = identify(AwDiagStatus::FATAL) ? AwDiagStatus::FATAL :
+    identify(AwDiagStatus::ERROR) ? AwDiagStatus::ERROR :
+    identify(AwDiagStatus::WARN) ? AwDiagStatus::WARN :
+    AwDiagStatus::OK;
+  new_status.type = AwDiagStatus::OUT_OF_RANGE;
+  return SET_DIAG_STATUS(new_status);
 }
 
-void HealthChecker::CHECK_RATE(std::string key, double warn_rate,
-                                     double error_rate, double fatal_rate,
-                                     std::string description) {
-  if (!keyExist(key)) {
-    std::unique_ptr<RateChecker> checker_ptr(
-      new RateChecker(autoware_health_checker::BUFFER_LENGTH, warn_rate,
-        error_rate, fatal_rate, description));
-    rate_checkers_[key] = std::move(checker_ptr);
+void HealthChecker::CHECK_RATE(const ErrorKey& key,
+  const double warn_rate, const double error_rate,
+  const double fatal_rate, const std::string& description)
+{
+  value_manager_.addCandidate(key);
+  if (value_manager_.isNotFound(key))
+  {
+    return;
   }
-  addNewBuffer(key, autoware_system_msgs::DiagnosticStatus::UNEXPECTED_RATE,
-               description);
+  if (!keyExist(key))
+  {
+    value_manager_.setDefaultValue(
+      key, "rate", warn_rate, error_rate, fatal_rate);
+    rate_checkers_[key] = std::make_unique<RateChecker>(
+      autoware_health_checker::BUFFER_DURATION,
+      warn_rate, error_rate, fatal_rate, description);
+  }
+  rate_checkers_[key]->setRate(
+    value_manager_.getValue(key, "rate", AwDiagStatus::WARN).get(),
+    value_manager_.getValue(key, "rate", AwDiagStatus::ERROR).get(),
+    value_manager_.getValue(key, "rate", AwDiagStatus::FATAL).get());
   rate_checkers_[key]->check();
+  addNewBuffer(key, AwDiagStatus::UNEXPECTED_RATE, description);
   return;
 }
+
+template <typename T>
+  autoware_system_msgs::DiagnosticStatus HealthChecker::setValueCommon(
+    const ErrorKey& key, const T& value, const std::string& desc)
+{
+  AwDiagStatus new_status;
+  new_status.header.stamp = ros::Time::now();
+  new_status.key = key;
+  new_status.value = valueToJson(value);
+  new_status.description = desc;
+  return new_status;
 }
+
+}  // namespace autoware_health_checker
