@@ -1,6 +1,3 @@
-#ifndef HEALTH_ANALYZER_H_INCLUDED
-#define HEALTH_ANALYZER_H_INCLUDED
-
 /*
  * Copyright 2019 Autoware Foundation. All rights reserved.
  *
@@ -20,18 +17,23 @@
  * v1.0 Masaya Kataoka
  */
 
+#ifndef AUTOWARE_HEALTH_CHECKER_HEALTH_ANALYZER_HEALTH_ANALYZER_H
+#define AUTOWARE_HEALTH_CHECKER_HEALTH_ANALYZER_HEALTH_ANALYZER_H
 // headers in ROS
 #include <ros/package.h>
 #include <ros/ros.h>
 
 // headers in STL
 #include <map>
+#include <string>
+#include <vector>
 
 // headers in Autoware
 #include <autoware_health_checker/constants.h>
 #include <autoware_system_msgs/SystemStatus.h>
 
 // headers in boost
+#include <boost/optional.hpp>
 #include <boost/foreach.hpp>
 #include <boost/graph/adjacency_list.hpp>
 #include <boost/graph/graph_utility.hpp>
@@ -40,59 +42,61 @@
 #include <boost/property_tree/ptree.hpp>
 #include <boost/property_tree/xml_parser.hpp>
 
-struct topic_property {
+struct topic_property
+{
   std::string node_pub;
   std::string node_sub;
 };
 
-struct node_property {
+struct node_property
+{
   std::string node_name;
 };
 
-typedef boost::adjacency_list<boost::listS, boost::vecS, boost::bidirectionalS,
-                              node_property, topic_property>
-    graph_t;
-typedef graph_t::vertex_descriptor vertex_t;
-typedef graph_t::edge_descriptor edge_t;
-typedef boost::graph_traits<graph_t>::adjacency_iterator adjacency_iterator_t;
-typedef boost::graph_traits<graph_t>::out_edge_iterator out_edge_iterator_t;
 
-class HealthAnalyzer {
+class HealthAnalyzer
+{
 public:
   HealthAnalyzer(ros::NodeHandle nh, ros::NodeHandle pnh);
   ~HealthAnalyzer();
 
 private:
+  using graph_t = boost::adjacency_list<boost::listS, boost::vecS,
+  boost::bidirectionalS, node_property, topic_property>;
+  using vertex_t = graph_t::vertex_descriptor;
+  using edge_t = graph_t::edge_descriptor;
+  using adjacency_iterator_t = boost::graph_traits<graph_t>::adjacency_iterator;
+  using out_edge_iterator_t = boost::graph_traits<graph_t>::out_edge_iterator;
+
+  using AwDiagStatus = autoware_system_msgs::DiagnosticStatus;
   ros::Subscriber system_status_sub_;
   ros::Subscriber topic_statistics_sub_;
   ros::Publisher system_status_summary_pub_;
   ros::NodeHandle nh_;
   ros::NodeHandle pnh_;
-  void
-  systemStatusCallback(const autoware_system_msgs::SystemStatus::ConstPtr msg);
-  void generateDependGraph(autoware_system_msgs::SystemStatus status);
-  void addDepend(rosgraph_msgs::TopicStatistics statistics);
-  autoware_system_msgs::SystemStatus
-  filterSystemStatus(autoware_system_msgs::SystemStatus status);
-  std::vector<std::string>
-  findWarningNodes(autoware_system_msgs::SystemStatus status);
-  std::vector<std::string>
-  findErrorNodes(autoware_system_msgs::SystemStatus status);
-  std::vector<std::string> findRootNodes(std::vector<std::string> target_nodes);
-  bool nodeAlreadyExistsInGraph(std::string target_node);
-  vertex_t getTargetNode(std::string target_node);
-  int countWarn(autoware_system_msgs::SystemStatus msg);
+  void systemStatusCallback(
+    const autoware_system_msgs::SystemStatus::ConstPtr& msg);
+  void generateDependGraph(const autoware_system_msgs::SystemStatus& status);
+  void addDepend(const rosgraph_msgs::TopicStatistics& statistics);
+  autoware_system_msgs::SystemStatus filterSystemStatus(
+    const autoware_system_msgs::SystemStatus& status);
+  std::vector<std::string> findWarningNodes(
+    const autoware_system_msgs::SystemStatus& status);
+  std::vector<std::string> findErrorNodes(
+    const autoware_system_msgs::SystemStatus& status);
+  std::vector<std::string> findRootNodes(
+    const std::vector<std::string>& target_nodes);
+  boost::optional<vertex_t> getTargetNode(const std::string& target_node);
+  int countWarn(const autoware_system_msgs::SystemStatus& msg);
   void writeDot();
   graph_t depend_graph_;
   int warn_nodes_count_threshold_;
-  template <typename T> bool isAlreadyExist(std::vector<T> vector, T target) {
-    for (auto itr = vector.begin(); itr != vector.end(); itr++) {
-      if (target == *itr) {
-        return true;
-      }
-    }
-    return false;
-  };
+  template <typename T> bool
+    isAlreadyExist(const std::vector<T>& vector, const T& target) const
+  {
+    auto result = std::find(vector.begin(), vector.end(), target);
+    return (result != vector.end());
+  }
 };
 
-#endif // HEALTH_ANALYZER_H_INCLUDED
+#endif  // AUTOWARE_HEALTH_CHECKER_HEALTH_ANALYZER_HEALTH_ANALYZER_H
