@@ -1,6 +1,3 @@
-#ifndef HEALTH_CHECKER_H_INCLUDED
-#define HEALTH_CHECKER_H_INCLUDED
-
 /*
  * Copyright 2019 Autoware Foundation. All rights reserved.
  *
@@ -20,6 +17,8 @@
  * v1.0 Masaya Kataoka
  */
 
+#ifndef AUTOWARE_HEALTH_CHECKER_HEALTH_CHECKER_HEALTH_CHECKER_H
+#define AUTOWARE_HEALTH_CHECKER_HEALTH_CHECKER_HEALTH_CHECKER_H
 // headers in ROS
 #include <ros/ros.h>
 
@@ -27,6 +26,7 @@
 #include <autoware_health_checker/constants.h>
 #include <autoware_health_checker/health_checker/diag_buffer.h>
 #include <autoware_health_checker/health_checker/rate_checker.h>
+#include <autoware_health_checker/health_checker/value_manager.h>
 #include <autoware_system_msgs/NodeStatus.h>
 
 // headers in STL
@@ -34,6 +34,9 @@
 #include <map>
 #include <memory>
 #include <sstream>
+#include <string>
+#include <vector>
+#include <utility>
 
 // headers in boost
 #include <boost/bind.hpp>
@@ -41,84 +44,99 @@
 #include <boost/optional.hpp>
 #include <boost/property_tree/json_parser.hpp>
 #include <boost/property_tree/ptree.hpp>
-#include <boost/thread.hpp>
 
-namespace autoware_health_checker {
-class HealthChecker {
+namespace autoware_health_checker
+{
+using MinMax = std::pair<double, double>;
+class HealthChecker
+{
 public:
+  using AwDiagStatus = autoware_system_msgs::DiagnosticStatus;
+  using AwDiagStatusArray = autoware_system_msgs::DiagnosticStatusArray;
   HealthChecker(ros::NodeHandle nh, ros::NodeHandle pnh);
-  ~HealthChecker();
   void ENABLE();
-  uint8_t CHECK_MIN_VALUE(std::string key, double value, double warn_value,
-                          double error_value, double fatal_value,
-                          std::string description);
-  uint8_t CHECK_MAX_VALUE(std::string key, double value, double warn_value,
-                          double error_value, double fatal_value,
-                          std::string description);
-  // std::pair<double,double> first value is min value and second value is max
-  // value
-  uint8_t CHECK_RANGE(std::string key, double value,
-                      std::pair<double, double> warn_value,
-                      std::pair<double, double> error_value,
-                      std::pair<double, double> fatal_value,
-                      std::string description);
-  template <class T>
-  uint8_t CHECK_VALUE(
-      std::string key, T value, std::function<uint8_t(T value)> check_func,
-      std::function<boost::property_tree::ptree(T value)> value_json_func,
-      std::string description) {
-    addNewBuffer(key, autoware_system_msgs::DiagnosticStatus::INVALID_VALUE,
-                 description);
-    uint8_t check_result = check_func(value);
+  ErrorLevel CHECK_MIN_VALUE(const ErrorKey& key, const double value,
+    const double warn_value, const double error_value,
+    const double fatal_value, const std::string& description);
+  ErrorLevel CHECK_MAX_VALUE(const ErrorKey& key, const double value,
+    const double warn_value, const double error_value,
+    const double fatal_value, const std::string& description);
+  ErrorLevel CHECK_RANGE(const ErrorKey& key, const double value,
+    const MinMax warn_value, const MinMax error_value,
+    const MinMax fatal_value, const std::string& description);
+  template <class T> ErrorLevel CHECK_VALUE(
+    const ErrorKey& key, const T& value,
+    std::function<ErrorLevel(T value)> check_func,
+    std::function<boost::property_tree::ptree(T value)> value_json_func,
+    const std::string& description)
+  {
+    value_manager_.addCandidate(key);
+    if (value_manager_.isNotFound(key))
+    {
+      return AwDiagStatus::UNDEFINED;
+    }
+    ErrorLevel check_result = check_func(value);
     boost::property_tree::ptree pt = value_json_func(value);
     std::stringstream ss;
     write_json(ss, pt);
-    autoware_system_msgs::DiagnosticStatus new_status;
-    new_status.type = autoware_system_msgs::DiagnosticStatus::INVALID_VALUE;
+    AwDiagStatus new_status;
+    new_status.key = key;
     new_status.level = check_result;
     new_status.description = description;
     new_status.value = ss.str();
     new_status.header.stamp = ros::Time::now();
-    diag_buffers_[key]->addDiag(new_status);
-    return new_status.level;
+    new_status.type = AwDiagStatus::INVALID_VALUE;
+    return SET_DIAG_STATUS(new_status);
   }
-  void CHECK_RATE(std::string key, double warn_rate, double error_rate,
-                  double fatal_rate, std::string description);
-  uint8_t CHECK_TRUE(std::string key, bool value, uint8_t level, std::string description);
-  uint8_t SET_DIAG_STATUS(autoware_system_msgs::DiagnosticStatus);
-  void NODE_ACTIVATE() {
+  void CHECK_RATE(const ErrorKey& key, const double warn_rate,
+    const double error_rate, const double fatal_rate,
+    const std::string& description);
+  ErrorLevel CHECK_TRUE(const ErrorKey& key, const bool value,
+    const ErrorLevel level, const std::string& description);
+  ErrorLevel SET_DIAG_STATUS(
+    const autoware_system_msgs::DiagnosticStatus& status);
+  void NODE_ACTIVATE()
+  {
     std::lock_guard<std::mutex> lock(mtx_);
     node_activated_ = true;
   };
-  void NODE_DEACTIVATE() {
+  void NODE_DEACTIVATE()
+  {
     std::lock_guard<std::mutex> lock(mtx_);
     node_activated_ = false;
   };
-  bool getNodeStatus() { return node_activated_; };
+  bool getNodeStatus()
+  {
+    return node_activated_;
+  };
 
 private:
-  std::vector<std::string> getKeys();
-  std::vector<std::string> getRateCheckerKeys();
+  std::vector<ErrorKey> getKeys();
+  std::vector<ErrorKey> getRateCheckerKeys();
   ros::NodeHandle nh_;
   ros::NodeHandle pnh_;
-  std::map<std::string, std::unique_ptr<DiagBuffer>> diag_buffers_;
-  std::map<std::string, std::unique_ptr<RateChecker>> rate_checkers_;
+  ros::Timer timer_;
+  ValueManager value_manager_;
+  std::map<ErrorKey, std::unique_ptr<DiagBuffer>> diag_buffers_;
+  std::map<ErrorKey, std::unique_ptr<RateChecker>> rate_checkers_;
   ros::Publisher status_pub_;
-  bool keyExist(std::string key);
-  bool addNewBuffer(std::string key, uint8_t type, std::string description);
-  template <typename T> std::string valueToJson(T value) {
-    using namespace boost::property_tree;
+  bool keyExist(const ErrorKey& key) const;
+  bool addNewBuffer(const ErrorKey& key, const ErrorType type,
+    const std::string& description);
+  template <typename T> autoware_system_msgs::DiagnosticStatus
+    setValueCommon(
+      const ErrorKey& key, const T& value, const std::string& desc);
+  template <typename T> std::string valueToJson(const T& value)
+  {
     std::stringstream ss;
-    ptree pt;
+    boost::property_tree::ptree pt;
     pt.put("value", value);
     write_json(ss, pt);
     return ss.str();
   }
-  void publishStatus();
+  void publishStatus(const ros::TimerEvent& event);
   bool node_activated_;
   std::mutex mtx_;
-  boost::thread publish_thread_;
-  bool ros_ok_;
 };
-}
-#endif // HEALTH_CHECKER_H_INCLUDED
+}  // namespace autoware_health_checker
+#endif  // AUTOWARE_HEALTH_CHECKER_HEALTH_CHECKER_HEALTH_CHECKER_H
