@@ -17,100 +17,116 @@
  * v1.0 Masaya Kataoka
  */
 
+#include <string>
+#include <vector>
+#include <algorithm>
 #include <autoware_health_checker/health_checker/diag_buffer.h>
 
-namespace autoware_health_checker {
-DiagBuffer::DiagBuffer(std::string key, uint8_t type, std::string description,
-                       double buffer_length)
-    : type(type), description(description) {
-  key_ = key;
-  buffer_length_ = ros::Duration(buffer_length);
-}
+namespace autoware_health_checker
+{
 
-DiagBuffer::~DiagBuffer() {}
+DiagBuffer::DiagBuffer(ErrorKey key, ErrorType type,
+  std::string description, double buffer_duration)
+  : type(type)
+  , description(description)
+  , key_(key)
+  , buffer_duration_(ros::Duration(buffer_duration)) {}
 
-void DiagBuffer::addDiag(autoware_system_msgs::DiagnosticStatus status) {
+void DiagBuffer::addDiag(autoware_system_msgs::DiagnosticStatus status)
+{
   std::lock_guard<std::mutex> lock(mtx_);
-  buffer_[status.level].status.emplace_back(status);
+  buffer_.at(status.level).status.emplace_back(status);
   updateBuffer();
   return;
 }
 
-autoware_system_msgs::DiagnosticStatusArray DiagBuffer::getAndClearData() {
+autoware_system_msgs::DiagnosticStatusArray DiagBuffer::getAndClearData()
+{
+  static const std::array<ErrorLevel, 4> level_array =
+  {
+    AwDiagStatus::ERROR,
+    AwDiagStatus::WARN,
+    AwDiagStatus::OK,
+    AwDiagStatus::UNDEFINED
+  };
   std::lock_guard<std::mutex> lock(mtx_);
-  autoware_system_msgs::DiagnosticStatusArray data;
-  data = buffer_[autoware_health_checker::LEVEL_FATAL];
-  data.status.insert(
-      data.status.end(),
-      buffer_[autoware_health_checker::LEVEL_ERROR].status.begin(),
-      buffer_[autoware_health_checker::LEVEL_ERROR].status.end());
-  data.status.insert(
-      data.status.end(),
-      buffer_[autoware_health_checker::LEVEL_WARN].status.begin(),
-      buffer_[autoware_health_checker::LEVEL_WARN].status.end());
-  data.status.insert(data.status.end(),
-                     buffer_[autoware_health_checker::LEVEL_OK].status.begin(),
-                     buffer_[autoware_health_checker::LEVEL_OK].status.end());
-  data.status.insert(
-      data.status.end(),
-      buffer_[autoware_health_checker::LEVEL_UNDEFINED].status.begin(),
-      buffer_[autoware_health_checker::LEVEL_UNDEFINED].status.end());
-  std::sort(data.status.begin(), data.status.end(),
-            std::bind(&DiagBuffer::isOlderTimestamp, this,
-                      std::placeholders::_1, std::placeholders::_2));
+  auto data = buffer_.at(AwDiagStatus::FATAL);
+  auto& status = data.status;
+  for (const auto& level : level_array)
+  {
+    const auto& buf_status = buffer_.at(level).status;
+    status.insert(status.end(), buf_status.begin(), buf_status.end());
+  }
+  std::sort(status.begin(), status.end(),
+    std::bind(&DiagBuffer::isOlderTimestamp, this,
+    std::placeholders::_1, std::placeholders::_2));
   buffer_.clear();
   return data;
 }
 
-uint8_t DiagBuffer::getErrorLevel() {
+ErrorLevel DiagBuffer::getErrorLevel()
+{
+  static const std::array<ErrorLevel, 4> level_array =
+  {
+    AwDiagStatus::FATAL,
+    AwDiagStatus::ERROR,
+    AwDiagStatus::WARN,
+    AwDiagStatus::OK
+  };
   std::lock_guard<std::mutex> lock(mtx_);
   updateBuffer();
-  if (buffer_[autoware_health_checker::LEVEL_FATAL].status.size() != 0) {
-    return autoware_health_checker::LEVEL_FATAL;
-  } else if (buffer_[autoware_health_checker::LEVEL_ERROR].status.size() != 0) {
-    return autoware_health_checker::LEVEL_ERROR;
-  } else if (buffer_[autoware_health_checker::LEVEL_WARN].status.size() != 0) {
-    return autoware_health_checker::LEVEL_WARN;
-  } else {
-    return autoware_health_checker::LEVEL_OK;
+  for (const auto& level : level_array)
+  {
+    if (!buffer_.at(level).status.empty())
+    {
+      return level;
+    }
   }
+  return AwDiagStatus::OK;
 }
 
 // filter data from timestamp and level
 autoware_system_msgs::DiagnosticStatusArray
-DiagBuffer::filterBuffer(ros::Time now, uint8_t level) {
-  autoware_system_msgs::DiagnosticStatusArray filterd_data;
-  autoware_system_msgs::DiagnosticStatusArray ret;
-  if (buffer_.count(level) != 0) {
-    filterd_data = buffer_[level];
+DiagBuffer::filterBuffer(ros::Time now, ErrorLevel level)
+{
+  autoware_system_msgs::DiagnosticStatusArray filtered_data;
+  if (buffer_.count(level) != 0)
+  {
+    filtered_data = buffer_.at(level);
   }
-  for (auto data_itr = filterd_data.status.begin();
-       data_itr != filterd_data.status.end(); data_itr++) {
-    if (data_itr->header.stamp > (now - buffer_length_)) {
-      ret.status.push_back(*data_itr);
+  autoware_system_msgs::DiagnosticStatusArray ret;
+  for (const auto& data : filtered_data.status)
+  {
+    if (data.header.stamp > (now - buffer_duration_))
+    {
+      ret.status.push_back(data);
     }
   }
   return ret;
 }
 
-void DiagBuffer::updateBuffer() {
+void DiagBuffer::updateBuffer()
+{
+  static const std::array<ErrorLevel, 5> level_array =
+  {
+    AwDiagStatus::FATAL,
+    AwDiagStatus::ERROR,
+    AwDiagStatus::WARN,
+    AwDiagStatus::OK,
+    AwDiagStatus::UNDEFINED
+  };
   ros::Time now = ros::Time::now();
-  buffer_[autoware_health_checker::LEVEL_FATAL] =
-      filterBuffer(now, autoware_health_checker::LEVEL_FATAL);
-  buffer_[autoware_health_checker::LEVEL_ERROR] =
-      filterBuffer(now, autoware_health_checker::LEVEL_ERROR);
-  buffer_[autoware_health_checker::LEVEL_WARN] =
-      filterBuffer(now, autoware_health_checker::LEVEL_WARN);
-  buffer_[autoware_health_checker::LEVEL_OK] =
-      filterBuffer(now, autoware_health_checker::LEVEL_OK);
-  buffer_[autoware_health_checker::LEVEL_UNDEFINED] =
-      filterBuffer(now, autoware_health_checker::LEVEL_UNDEFINED);
+  for (const auto& level : level_array)
+  {
+    buffer_[level] = filterBuffer(now, level);
+  }
   return;
 }
 
 bool DiagBuffer::isOlderTimestamp(
     const autoware_system_msgs::DiagnosticStatus &a,
-    const autoware_system_msgs::DiagnosticStatus &b) {
+    const autoware_system_msgs::DiagnosticStatus &b)
+{
   return a.header.stamp < b.header.stamp;
 }
-}
+}  // namespace autoware_health_checker
