@@ -19,35 +19,33 @@
 
 #include <autoware_health_checker/system_status_subscriber/system_status_subscriber.h>
 
-namespace autoware_health_checker {
-SystemStatusSubscriber::SystemStatusSubscriber(ros::NodeHandle nh,
-                                               ros::NodeHandle pnh) {
-  nh_ = nh;
-  pnh_ = pnh;
-}
-
-SystemStatusSubscriber::~SystemStatusSubscriber() {}
-
-void SystemStatusSubscriber::enable() {
-  status_sub_ = nh_.subscribe("system_status", 10, &SystemStatusSubscriber::systemStatusCallback, this);
-}
-
-void SystemStatusSubscriber::systemStatusCallback(const autoware_system_msgs::SystemStatus::ConstPtr msg)
+namespace autoware_health_checker
 {
-  for (auto function_itr = functions_.begin(); function_itr != functions_.end(); function_itr++)
+SystemStatusSubscriber::SystemStatusSubscriber(
+  ros::NodeHandle nh, ros::NodeHandle pnh)
+  : nh_(nh), pnh_(pnh) {}
+
+void SystemStatusSubscriber::enable()
+{
+  status_sub_ = nh_.subscribe(
+    "system_status", 10, &SystemStatusSubscriber::systemStatusCallback, this);
+}
+
+void SystemStatusSubscriber::systemStatusCallback(
+  const autoware_system_msgs::SystemStatus::ConstPtr msg)
+{
+  for (auto& func : functions_)
   {
-    std::function<void(std::shared_ptr<autoware_system_msgs::SystemStatus>)> func = *function_itr;
-    std::shared_ptr<autoware_system_msgs::SystemStatus> pmsg =
-    std::make_shared<autoware_system_msgs::SystemStatus>(*msg);
+    auto pmsg = std::make_shared<autoware_system_msgs::SystemStatus>(*msg);
     func(pmsg);
   }
   return;
 }
 
-void SystemStatusSubscriber::addCallback(std::function<void(std::shared_ptr<autoware_system_msgs::SystemStatus>)> func)
+void SystemStatusSubscriber::addCallback(
+  std::function<void(std::shared_ptr<autoware_system_msgs::SystemStatus>)> func)
 {
   functions_.push_back(func);
   return;
 }
-
-}
+}  // namespace autoware_health_checker
